+++ src/lib.rs
@@ -69,6 +69,6 @@ pub fn app<M: 'static>(model: app::ModelFn<M>) -> app::Builder<M, Event> {
 ///
 /// This is useful for late night hack sessions where you just don't care about all that other
 /// stuff, you just want to play around with some ideas or make something pretty.
-pub fn sketch(view: app::SketchViewFn) {
+pub fn sketch(view: app::SketchViewFn) -> app::SketchBuilder<Event> {
     app::Builder::sketch(view)
 }